not working

+++ tests/datatools/test_wtseries_tensor.py
@@ -19,6 +19,6 @@ def test_WTSeriesTensor():
     dataset = WTSeriesTensor(TRAIN, ts_data=ts_data)
     for i in range(len(dataset)):
         inp, tar = dataset[i]
-        assert np.all(np.array(inp) == np.array(ts_data(TRAIN, INPUT)[i]))
-        assert np.all(np.array(tar) == np.array(ts_data(TRAIN, TARGET)[i]))
+        assert np.all(np.array(inp) == np.array(ts_data(TRAIN, INPUT)[i].astype(np.float32)))
+        assert np.all(np.array(tar) == np.array(ts_data(TRAIN, TARGET)[i].astype(np.float32)))
     assert dataset.input_data.size() == ts_data(TRAIN, INPUT).shape()
